pyproject.toml
[tool.poetry]
name = "transcricao-video-ia"
version = "0.1.0"
description = ""
authors = ["mielesantos <mielesnts@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
langchain-community = "^0.3.7"
langchain-huggingface = "^0.1.2"
langchain-ollama = "^0.2.0"
langchain-openai = "^0.2.9"
langchain-groq = "^0.2.1"
pytube = "^15.0.0"
python-dotenv = "^1.0.1"

[tool.poetry.group.dev.dependencies]
taskipy = "^1.12.2"
ruff = "^0.4.8"
httpx = "^0.27.0"

[tool.ruff]
line-length = 88
extend-exclude = ['migrations']

[tool.ruff.lint]
preview = true
select = ['I', 'F', 'E', 'W', 'PL', 'PT']

[tool.ruff.format]
preview = true
quote-style = 'single'

[tool.pytest.ini_options]
pythonpath = "."
addopts = '-p no:warnings'

[tool.taskipy.tasks]
lint = 'ruff check .&& ruff check . --diff'
format = 'ruff check . --fix; ruff format .'
run = 'python app/main.py'
pre_test = 'task lint'

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"